+++ src/XRP/reliable-submission-xrp-client.ts
@@ -144,3 +144,4 @@ export default class ReliableSubmissionXrpClient implements XrpClientDecorator {
     return this.decoratedClient.getPayment(transactionHash)
   }
 }
+