+++ src/XRP/xrp-client-decorator.ts
@@ -116,3 +116,4 @@ export default interface XrpClientDecorator {
    */
   getPayment(transactionHash: string): Promise<XrpTransaction | undefined>
 }
+