+++ src/index.ts
@@ -15,8 +15,7 @@ export { default as XpringClient } from './Xpring/xpring-client'
 export { default as XrpError } from './XRP/xrp-error'
 export { default as IlpError, IlpErrorType } from './ILP/ilp-error'
 export { default as XrpPayIdClient } from './PayID/xrp-pay-id-client'
-export { XrpUtils } from './XRP'
-export { RippledFlags, TransactionStatus, XrpClient } from './XRP'
+export { RippledFlags, TransactionStatus, XrpClient, XrpUtils } from './XRP'
 export {
   SendXrpDetails,
   XrpCurrencyAmount,