+++ CHANGELOG.md
@@ -31,7 +31,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `XRPError` was deprecated for two releases and has been removed. Use `XrpError` instead.
 - `XRPErrorType` was deprecated for two releases and has been removed. Use `XrpErrorType` instead.
 - `XrpTransaction.account` and `XrpTransaction.sourceTag` were deprecated for two releases and have been removed.
-  Please use the X-address encoded field `sourceXAddress` instead.
+  Use the X-address encoded field `sourceXAddress` instead.
 - `XrpPayment.destination` and `XrpPayment.destinationTag` were deprecated for two releases and have been removed.
   Please use the X-address encoded field `destinationXAddress` instead.